make reuse work on Windows
