+++ netlink_other.go
@@ -1,8 +1,9 @@
 // +build !linux
+// +build !windows
 
 package libp2pquic
 
 import "github.com/vishvananda/netlink/nl"
 
-// nl.SupportedNlFamilies is the default netlink families used by the netlink package
+// SupportedNlFamilies is the default netlink families used by the netlink package
 var SupportedNlFamilies = nl.SupportedNlFamilies
